From String to bytes!
Starting with read_to_string was great. but it was time to move to just reading the raw bytes - in hopes to be able to handle more than just plain text files (think images and whatnot)
so some major refactoring went into motion.

created a new struct and enum to help with this restructure. specifically so that all the different pieces of an http response could be grouped.
Needed some new logic for figuring out how to deconstruct the struct and enum back into pieces so I could write that to the stream.
I wish I could give more details but this is already kinda long - see code for how I did that